Limit results and alias the values
Aliasing the values because the front-end works much easier with
them. This could be done in template code, or the charting function
could be changed instead.

Limits the results by setting per page and then 1st page. Confused
because I had to multiply the per page by 2, so will look at this
after.