Added time label to bottom of chart and modified chart layout
Question:
Here, as per the mockup, I modified the chart so only the top "tick" or
number is showing. Is this supposed to be done for all charts to keep
consistency? Or am I supposed to write special code for just this chart.

Also, I may be confused, but I believe the issue says to take the last
4 reports, but the mockup has 9 separate reports showing. Which is
correct? This is making the width of the chart different from the
mockups.