Changed color of charts
I need to talk to Noah about the pattern he wanted implemented in
the mockup, this is close to that color though.

I also was unsure of color organization. I created two new color vars
for this, but I think it would've been better to try and reuse existing
colors. Am unsure of which ones to use though.