Styling changes for tables
I modified the spacing of the table names to the right so they are
centered. This method doesn't support if they two lines of text.

I changed the borders so they are below rather then above. There is
still some more work to do here that I wanted to save for when I make
the acoordion as it will affect it.