Don't use deprecated mixin.
Bourbon is deprecating the `box-sizing` mixin because the CSS property
is now well-supported across browsers. Using the mixin throws a warning
with the current version of Bourbon and will error after the next major
release. If we want to support very old browser versions, it's safest to
add `-moz-box-sizing` etc. manually, but this probably isn't necessary.