Fix grouping in main typeahead search.
The actual issue here turned out to be that we were checking the
candidate and committee IDs against `null` but were getting back a
different falsy value of `''`. This patch uses a more general check,
ignoring results with all falsy ID values, and revises the filtering
logic to use a more explicit filter and map, rather than the implicit
filtering in `$.map`.

[Resolves https://github.com/18F/openFEC/issues/741]