Use full-text search for committee and candidate search fields.
This patch passes the name search value as "q" instead of "name",
meaning that we use the full-text search on candidate and committee
names instead of the stricter `ILIKE` used by "name". Practically, this
means that searching for "nancy pelosi", "pelosi nancy", or "pelosi,
nancy" will all yield Nancy Pelosi, which is what I think users will
expect.

One side effect of this proposal is that searching for a
candidate's FEC ID will also find that candidate, since the full-text
search includes both name and ID. If this is a problem, we can fix it on
the backend, but I don't think it's a cause for concern.

[Resolves https://github.com/18F/openFEC/issues/777]