Show all cycles between first and last election.
Currently, the election cycle picker for candidates only shows the
election cycles for which the candidate has run. This can be confusing
for Senate and Presidential candidates, since financial reports for
off-election cycles may also be relevant. This patch updates the
candidate cycle picker to show all cycles between the candidate's first
and last elections.

Note: The implementation of this change requires that we distinguish
between the cycle used to find associated committees and reports versus
the cycle used to look up candidate history. For off-election cycles,
the year used to look up the candidate history is passed through the
query string as `history`. IMO this is a bit of a kludge, but I'd rather
have the logic in the webapp than change the API for an edge case like
this. I'm open to alternatives here.

h/t @noahmanger for suggesting