Generate correct history parameter when next cycle is not available.
We use the `next_cycle` filter to generate the `history` query parameter
for the candidate detail page. The filter is meant to get the next
active cycle greater or equal to the specified cycle. This patch handles
the edge case in which no such cycle is found, in which case we should
simply use the latest active cycle. This fixes some errors on Jeff
Flake's candidate page.