User wants to see most accurate search suggestions
Changing typeahead implementation so it switches the data source
based on whether candidate or committee was selected so that the user
only gets suggestions based on what they have selected.

Before there was no selector so it was showing both.

TODO: we may want to change the limit from 5 to 10 as there's more room
now, @noahmanger.