Move HTML markup for "None" to macro.
Putting HTML in a string literal in the `currency` filter was breaking
tests. It's also not ideal, since it requires using the `safe` filter,
and unsafe content could potentially leak through. This patch restores
the behavior of the `currency` filter and moves the markup for the "None"
element to a macro.