Improve filter button
This moves the filter button to the right side of the filter bar. This
was done so the filter button UI was attached to the actual filter
UI, so the button affords closing and opening the side bar.

This iteration doesn't include any animation. Animation would require
some extra work.