Attempt to fix filter panel
I changed the states filter macro to a partial again. This was
because my filter container needs partials to work. This still
has to be fixed on the expeditures and donations pages. @jmcarp,
let me know if you kow better ways to do this with jinja.

I also fixed the positioning of the filter on mobile with a little
hack.