Getting panel to work on modal
- Rename the UI of the pop-out from "modal" to "panel". This was renamed
because this UI will not be strictly a modal. It will act as a modal
on mobile and as a split 50% panel on tablet, up. @noahmanger, please
let me know if theres a better name.
- Change the columns around to have an arrow column at the end of the
rows.
- Make the whole row clickable to open the panel, if you click on a link
it will pass through and open the link.

This needs some of the work done in #337, #338 in order to get
expeditures working.