Large screen indentation of tds and headers
The headers are supposed to be smaller on larger screens. There was
an indentation problem with the fields of the panel. This was due
to them being a td under dataTable, which in turn uses an !important
rule likely because it has to override dataTable's inline code.

This all has to be fixed but would likely be a larger refactor,
so am ignoring now with a shameful rule.