Arrow button on rows for panel
This UI is to make it easier to understand how to open the table
rows to show the details panel. A new table column was used to
contain the button so that everything fits within data tables
width interface. Additional styles were set from the table scss
as its generally safe to ensure these styles for any icon in the
table.

@noahmanger whats the best way to make the arrows as thick as in
the mockup? Right now they are thinner.