Fixed data tables half width
The problem was that data tables uses absolute px values to calculate
the table columns, which means if you set the table to be 50%, it will
not work like that.

This changes the table so it's contained in a div which is 100% and
switches to 50% when the panel is active. From here, we have to ensure
data tables hides the columns which are supposed to be hidden for
the panel. This is done manually in JS (it can't be done in CSS or
data tables won't calculate the other columns correctly).

It seems this gets the data tables working on all breakpoints so its
split 50/50 with the panel when it's active.