Only show table after it's finished drawing
Only applies to the first time the page is loaded. This prevents the
jumbled table headers from showing up on mobile before the table is
loaded, and i think makes for a nice effect in general.

In fed-style I set the opacity to fade in. What do you think @jmcarp