Fix datetime formatter options.
The current datetime formatting logic works, but only because of a quirk
of the javascript interpreter. When passing key-value pairs to
handlebars helpers, handlebars passes an options object to the helper
function; the options are collected in the `hash` attribute of the
options object. So the signature of the helper function needs to be
`(value, options)`, and the function needs to check `options.hash` for
option values. The current code uses the wrong signature, but it works
because we only ever pass one option, and objects are truthy in
javascript.

@noahmanger #themoreyouknow