Adding links to forms that have the relevant information for the Presidential and independent expenditure totals.
We don't show the IE spending on that page anymore, but
There isn't a better way to explain where we get that
number. Not having a link for it might imply that the
presidential link would give the information, which it
doesn't.

We could improve this in the future, when we have a
view for IEs.

resolves #697