Replace history on table initialization.
This patch attempts to fix a longtime issue with permalinks for table
filters. On table initialization, we grab filter values from the query
string and set them in the filter fields. Because some fields change
their inputs (mostly fields with input masks, but potentially other
fields as well), this can trigger a change in the URL. In these cases,
the back button brings the user back to the initial URL, which again
triggers an immediate URL change when the input is changed by the filter
field. This patch sidesteps the issue by using `history.replaceState`
rather than `history.pushState` on load, meaning that we don't push a
new entry to this history.

To verify, enter a URL that will be automatically transformed on load,
such as http://localhost:3000/filings/?min_receipt_date=10/01/1985, and
verify that the back button works as expected.

cc @noahmanger