Fixes the app initialization with gunicorn
This changeset addresses an issue I introduced with the last commit to update the deployment.  I tried to account for this by explicitly setting the app name to `openfecwebapp` in the Flask app initialization, but by moving it into `app.py` this still ultimately did not work.  I have changed this to simply use the module name itself as is standard practice and fixed the `gunicorn` initialization in `run.sh` to properly account for the new namespace.