Fix initial data table fetch.
As described in #1192, a recent refactor to data tables caused a
regression such that filters in query parameters were ignored on initial
fetch. This happened because the `DataTable` class was instantiated
before the `filterSet` attribute was set; since data tables depend on
the `filterSet` attribute to create filters, and since fetching starts
as soon as the `DataTable` is instantiated, the `filterSet` attribute
must be set first. This patch restores the order of operations and adds
a comment explaining why it matters.

[Resolves #1192]