Copy correction
cc @noahmanger

Aligning with @emileighoutlaw  original changes. If we can use
variables:

Sorry, we didn't find any financial data for **[this committee]** in
the **[selected two-year cycle]** election cycle.

Ideally we want it to read something like:
Sorry, we didn't find any financial data for ABATE, FRANK G in the
1991–1992 election cycle.