Removing cycle from links to committee pages from candidate page
This caused an error when, for instance, viewing the full cycle on a
candidate page, and there is a principal campaign committee from a
previous election shown. This code included the cycle, which lead to a
404.

Not sure if this is how we ultimately want to deal with it, but in the
mean time it removes breaking functionality.

cc @LindsayYoung