replacing pushQuery with updateQuery
Doing this because with the new transaction period filters, we need to
set filters by default if the page is loaded without them. This breaks
the ability to hit back to leave the page because pushQuery fires
repeatedly.

With this change, you can no longer use “back” to move between queries;
instead, “back” tacks you to the last distinct page you were on.