Fixes issue with undefined sort parameter
This changeset addresses the issue with the undefined sort parameter being added to the API calls when viewing the Disbursements tab on a single Committee page.  Note that we also had to excplicitly set the purpose value to be lower case for this to work.

h/t to @noahmanger for the huge assist in helping track down where this was being constructed and the right values to set.