Adding front-end support for downloading sched A + B aggregates
This adds a new configuration option to the DataTables class for
`aggregateExport`. The `useExport` option that the main data tables use
is used to configure several other components as part of the overall
export pattern (like tooltips), which we don’t need here.

If this parameter is passed in initialization, then the DataTable will
look for a button with the data-export-for attribute that matches its
data-type attribute. It’s a little circuitous, but because the export
button lives outside the `this.$body` of the table, it seemed like the
best bet.

It also cleans up the DataTable code by pulling out the tooltip init /
destroy code into separate methods.