Restoring pdf_url column
This adds back the `pdf_url` column which was replaced by
`document_type` which includes the button to download different
versions of a report. This is needed by the committee page filings
table, which is currently too narrow to support the button. This fixes
an error in production, but we’ll want to come up with a better
solution later.