Refactoring, adding documentation and tests
This breaks up several of the larger functions into smaller pieces that
are easier to test and refines some of the behavior around navigating
to future and past cycles.

It also fixes a bug where the dates on the chart were
timecode-adjusted, so they were showing up as a day earlier than they
should be, so I introduced a new helper method for converting to a UTC
date.