Updating tooling for web app
This changeset updates some of the tooling and infrastructure in the web app to bring things up to date.  It includes the following:

* Cleanup of .travis.yml
* Update fec-style to latest release (7.7.0)
* Update of some Python dependencies (invoke, py.test, requests)
* Clean up of tasks

The Travis cleanup includes updating nvm to make use of the latest stable release of Node.js, which I have been running locally for a while now with no issues.