Add dataType to ajax request
This changeset addresses the jQuery vulnerability we encountered a bit further by ensuring that a dataType parameter is set on our ajax requests.  We cannot do an upgrade to the 3.x series of jQuery just yet, so to help mitigate any potential threat we are ensuring our requests do not accept scripts to be executed.  We will update jQuery to the latest release in the near future after we have had a bit more time to test the changes and address any compatability issues.