Update gevent and revert log level
This changeset is a follow-up to our previous one to ensure we are using the latest stable version of gevent and to revert the log level setting for gunicorn.  We had originally modified it to gain a bit more insight into the performance issues we were having and see if there were any trends or patterns we could identify from the web app side.

h/t to @vrajmohan for all of the help!